+++ epyk/core/css/Colors.py
@@ -326,6 +326,10 @@ def colors(start, end, steps):
 
   :return: A list of hexadecimal color codes.
   """
+  if start.upper() in defined:
+    start = defined[start.upper()]['hex']
+  if end.upper() in defined:
+    start = defined[end.upper()]['hex']
   colors_panel = [start]
   for i in range(steps-2):
     colors_panel.append(gradient(start, end, 1.0 / (steps-1) * (i + 1)))