+++ epyk/core/html/geo/__init__.py
@@ -4,4 +4,5 @@ from . import GeoDc
 from . import GeoChartJs
 from . import GeoGoogle
 from . import GeoJqv
+from . import GeoD3
 from . import mappings